libta: refactored code
Code Refactoring of the library. Now libta.h is in include/libta folder.
To use the library just add the include folder in the include path.
Everything seems to work. The only issue seems that with too much points
in the rank the computation goes wrong and for small probabilities,
wrong execution time estimations are produced.